test: Add fault injection hooks and comprehensive tests
- Implement `lq_mem_simulate_oom` and `lq_io_simulate_error` hooks.
- Update `src/store/file.c` to use `lq_write` and safely check `lq_read` return values.
- Fix signature mismatch for `lq_read`/`lq_write` in `src/lq/io.h` (use `size_t`).
- Add `src/test/test_faults.c` covering basic faults and `src/store/file.c` error paths.
- Add `TODO.md` for coverage goals.

+++ src/lq/io.h
@@ -18,8 +18,8 @@
 char* mktempdir(char *s);
 char* ensuredir(char *s);
 int lq_open(const char *pathname, int flags, int mode);
-int lq_read(int f, char *buf, int c);
-int lq_write(int f, char *buf, int c);
+int lq_read(int f, char *buf, size_t c);
+int lq_write(int f, char *buf, size_t c);
 
 /**
  * @brief Configure I/O error simulation.
